requirements.txt
huggingface_hub[cli,torch]
hf_transfer
pillow
tensorboard
termcolor
torch
torchvision
torchaudio
tqdm